Retire the bundled-translations machinery, which now holds nothing
With Spanish loading like every other language, `BUNDLED_TRANSLATIONS` is an
always-empty map, `pick` only ever picks English, and the worker's
`BUNDLED_WORKER_RESOURCES` is an empty object spread into every translator it
builds. What was left read as though something might be in it when nothing ever
is, so `bundled.ts` is gone.

`BUNDLED_LOCALES` becomes the constant `["en"]` in `catalogs.ts`, beside the
English catalogs it now describes, and `bundledResources` becomes
`englishResources(namespaces)` — the same combine, named for what it returns.
The chrome keeps its map because it negotiates against those keys and needs `en`
as a candidate; the worker is handed nothing at all, English already being
appended behind every chain by `createTranslator`.

Inlining a locale again is now a deliberate change with a case to make — add it
to `BUNDLED_LOCALES`, exclude it from the lazy glob, give the chrome and the
worker a way to reach it — rather than filling in a seam left standing open.
`lint:i18n` still holds the glob's exclusions to `BUNDLED_LOCALES`, so the two
cannot drift.

Also swept the prose the unbundling left stale: the i18n, standalone and iframe
READMEs still said English and Spanish were both carried, `supportedLocales`'
generated header pointed at `bundledResources` for what the bundle holds, and
two comments in the bundle-size guard and one in a viewer spec named Spanish as
an inlined locale. The guard's multi-inlined-locale test now says it supplies
that second locale by hand rather than describing one that exists.
